Working SH shader

+++ Src/MathUtils.h
@@ -1,8 +1,8 @@
 #pragma once
 
-constexpr double sqrt2 = 1.41421356237309504880;
+constexpr float sqrt2 = 1.41421356237309504880;
 
-constexpr double factorial[35] = {
+constexpr float factorial[35] = {
     1.0,
     1.0,
     2.0,
